+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.4.0 - 2023-01-15
+### Changed
+- Prompt for creation of new config file with `-init`.
+### Added
+- `-force` option (to force `-init` without prompting).
+
 ## 1.3.0 - 2023-01-14
 ### Added
 - Show `(DOES NOT EXIST)` for paths that do not exist if in `-no-color` mode.